feat: update ad components and add scroll button, improve layout responsiveness
