+++ components/AdSlot.tsx
@@ -142,7 +142,7 @@ export default function AdSlot({
           minHeight: "100px",
           ...style
         }}
-        data-ad-client="ca-pub-6655505252648648"
+        data-ad-client="ca-pub-3920325569173233"
         data-ad-slot={slotId}
         data-ad-format={format}
         data-full-width-responsive={fullWidthResponsive ? "true" : "false"}
